Tests: Co-locate SwizzleCache baseline helpers with SwizzlerTestCase
